Unify code for writing results
