+++ src/Microsoft.DotNet.XHarness.TestRunners.Xunit/WasmThreadedTestRunner.cs
@@ -16,12 +16,5 @@ internal class WasmThreadedTestRunner : XUnitTestRunner
     }
 
     public override void WriteResultsToFile(TextWriter writer, XmlResultJargon jargon)
-    {
-        using var ms = new MemoryStream();
-        AssembliesElement.Save(ms);
-        ms.TryGetBuffer(out var bytes);
-        var base64 = Convert.ToBase64String(bytes, Base64FormattingOptions.None);
-        Console.WriteLine($"STARTRESULTXML {bytes.Count} {base64} ENDRESULTXML");
-        Console.WriteLine($"Finished writing {bytes.Count} bytes of RESULTXML");
-    }
+        => WasmXmlResultWriter.WriteOnSingleLine(AssembliesElement);
 }
